Move RestError into its own file
